fixed admin-page
