+++ src/com/gesz/controller/AdminLoginController.java
@@ -21,20 +21,17 @@ private static final long serialVersionUID = -3435554487273689111L;
 		
 		String username = request.getParameter("username");
 		String password = request.getParameter("password");
-		RequestDispatcher dispatcher = null;
 		
 		Authenticator authenticator = new Authenticator();
 		String result = authenticator.authenticate(username,password);
 		if(result.equals("success")) {
-			dispatcher = request.getRequestDispatcher("pages/adminControl.jsp");
 			User user = new User(username,password);
 			request.setAttribute("user", user);
 			HttpSession session = request.getSession();
 			session.setAttribute("user", user);
+			session.setAttribute("isAdmin", "true");
 		}else {
 			request.setAttribute("loginMessage", "Login failed, please try again");
-			dispatcher=request.getRequestDispatcher("pages/adminLogin.jsp");
 		}
-		dispatcher.forward(request, response);
 	}
 }
